+++ CHANGELOG.md
@@ -15,7 +15,9 @@ Increment the:
 
 ## [Unreleased]
 
-* [SDK] Add ComposableSampler interface and CompositeSampler (work in progress)
+* [SDK] Add ComposableSampler and CompositeSampler with the consistent
+  probability sampling variants (AlwaysOn, AlwaysOff, TraceIdRatioBased,
+  ParentThreshold, RuleBased)
   [#4028](https://github.com/open-telemetry/opentelemetry-cpp/issues/4028)
 
 * [RELEASE] Bump main branch to 1.28.0-dev